# backend/api/openapi.yaml
openapi: 3.1.0
info:
  title: PortalDots API
  version: 0.1.0
servers:
  - url: http://127.0.0.1:8080/v1
paths:
...
  /staff/circles/{circleID}:
    get:
      operationId: getStaffCircle
      summary: Return a circle editable from staff mode
      parameters:
        - in: path
          name: circleID
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Staff circle detail
          content:
            application/json:
              schema:
                type: object
                required: [id, name, nameYomi, groupName, groupNameYomi, participationTypeId, participationTypeName, tags, notes, status, statusReason, places]
                properties:
                  id:
                    type: string
                  name:
                    type: string
                  nameYomi:
                    type: string
                  groupName:
                    type: string
                  groupNameYomi:
                    type: string
                  participationTypeId:
                    type: string
                  participationTypeName:
                    type: string
                  tags:
                    type: array
                    items:
                      type: string
                  notes:
                    type: string
                  submittedAt:
                    type: string
                    nullable: true
                  status:
                    type: string
                  statusReason:
                    type: string
                  statusSetAt:
                    type: string
                    nullable: true
                  statusSetById:
                    type: string
                    nullable: true
                  places:
                    type: array
                    items:
                      type: string
        "401":
          description: Unauthenticated
        "403":
          description: Staff verification or circle admin role is required
        "404":
          description: Circle not found
    delete:
      operationId: deleteStaffCircle
      summary: Delete a circle from staff mode
      parameters:
        - in: path
          name: circleID
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthenticated
        "403":
          description: Staff verification or circle admin role is required
        "404":
          description: Circle not found
    put:
      operationId: putStaffCircle
      summary: Update a circle from staff mode
      parameters:
        - in: path
          name: circleID
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, nameYomi, groupName, groupNameYomi, participationTypeId]
              properties:
                name:
                  type: string
                  minLength: 1
                nameYomi:
                  type: string
                groupName:
                  type: string
                  minLength: 1
                groupNameYomi:
                  type: string
                participationTypeId:
                  type: string
                  minLength: 1
                notes:
                  type: string
                status:
                  type: string
                  enum: [pending, approved, rejected]
                statusReason:
                  type: string
                placeIds:
                  type: array
                  items:
                    type: string
      responses:
        "200":
          description: Updated circle
          content:
            application/json:
              schema:
                type: object
                required: [id, name, nameYomi, groupName, groupNameYomi, participationTypeId, participationTypeName, tags, notes, status, statusReason, places]
                properties:
                  id:
                    type: string
                  name:
                    type: string
                  nameYomi:
                    type: string
                  groupName:
                    type: string
                  groupNameYomi:
                    type: string
                  participationTypeId:
                    type: string
                  participationTypeName:
                    type: string
                  tags:
                    type: array
                    items:
                      type: string
                  notes:
                    type: string
                  submittedAt:
                    type: string
                    nullable: true
                  status:
                    type: string
                  statusReason:
                    type: string
                  statusSetAt:
                    type: string
                    nullable: true
                  statusSetById:
                    type: string
                    nullable: true
                  places:
                    type: array
                    items:
                      type: string
        "401":
          description: Unauthenticated
        "403":
          description: Staff verification or circle admin role is required
        "404":
          description: Circle not found
        "422":
          description: Validation error
...
  /staff/circles/{circleID}/members/{userID}:
    delete:
      operationId: deleteStaffCircleMember
      summary: Remove a member from a circle in staff mode
      parameters:
        - in: path
          name: circleID
          required: true
          schema:
            type: string
        - in: path
          name: userID
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Member removed
        "401":
          description: Unauthenticated
        "403":
          description: Staff verification or circle admin role is required
        "404":
          description: Circle or user not found
        "422":
          description: Validation error
...